Fix ApplicationLaunchProperty 2
